+++ js/src/integrations-manager.js
@@ -17,6 +17,15 @@ class IntegrationsManager {
   register(name, integration) {
     this.integrations[name] = integration;
   }
+
+  runNodeIntegration(data) {
+    let result = '';
+    const { nodeRun } = this.get(data.integrationName);
+    if (typeof(nodeRun) === 'function') {
+      result = nodeRun(data.payload);
+    }
+    return result;
+  }
 }
 
 export default new IntegrationsManager;