+++ lib/react_webpack_rails.rb
@@ -1,3 +1,9 @@
 require 'react_webpack_rails/version'
-require 'react_webpack_rails/engine' if defined?(Rails)
-require 'react_webpack_rails/railtie' if defined?(Rails)
+if defined?(Rails)
+  require 'react_webpack_rails/engine'
+  require 'react_webpack_rails/railtie'
+  require 'react_webpack_rails/node_helpers'
+  require 'react_webpack_rails/node_integration_runner'
+  require 'react_webpack_rails/errors/base'
+  require 'react_webpack_rails/errors/node_failure'
+end