+++ lib/react_webpack_rails/railtie.rb
@@ -2,9 +2,16 @@ require 'react_webpack_rails/view_helpers'
 
 module ReactWebpackRails
   class Railtie < ::Rails::Railtie
+    # Sensible defaults. Can be overridden in application.rb.
+
+    # gem related config default:
+    config.rwr = ActiveSupport::OrderedOptions.new
+    config.rwr.node_server_host = 'http://localhost:8081/'
+
+    # react integration related config default:
     config.react = ActiveSupport::OrderedOptions.new
-    # Sensible defaults. Can be overridden in application.rb
-    config.react.camelize_props = false # pass in an underscored hash but get a camelized hash
+    config.react.camelize_props = false
+    config.react.server_side = false
     config.react.shared = {}
 
     initializer 'react_webpack_rails.view_helpers' do