+++ lib/react_webpack_rails/view_helpers.rb
@@ -11,9 +11,17 @@ module ReactWebpackRails
       content_tag(html_tag, '', html_options, &block)
     end
 
-    def react_component(name, props = {}, options = { ssr: false })
+    def react_component(name, props = {}, options = {})
+      props = props.as_json
       props = camelize_props_key(props) if Rails.application.config.react.camelize_props
-      react_element('react-component', { props: props, name: name }, options)
+      if server_side(options.delete(:server_side))
+        result = NodeIntegrationRunner.new('react-component', props: props, name: name).run
+        react_element('react-component', { props: props, name: name }, options) do
+          result.html_safe
+        end
+      else
+        react_element('react-component', { props: props, name: name }, options)
+      end
     end
 
     def react_router(name)
@@ -28,6 +36,10 @@ module ReactWebpackRails
         h[k.to_s.camelize(:lower)] = v.is_a?(Hash) ? camelize_props_key(v) : v
         h
       end
+    end
+
+    def server_side(server_side)
+      server_side.nil? ? Rails.application.config.react.server_side : server_side
     end
   end
 end