+++ spec/rails4_dummy_app/app/assets/javascripts/application.js
@@ -15,4 +15,5 @@
 //= require turbolinks
 //= require react_integration
 //= require react_bundle
+
 //= require_tree .